Attaced classes to AD.Story namespace
