+++ Assets/Scripts/Editor/Controllers/EntryController.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using UnityEngine;
+using AD.Story;
 
 namespace StoryEditor.Controllers
 {
