+++ Assets/Scripts/Editor/EditorStoryBlock.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
+using AD.Story;
 
 namespace StoryEditor
 {
