+++ Assets/Scripts/Editor/StoryScriptEditorWindow.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using UnityEditor;
+using AD.Story;
 using StoryEditor.Controllers;
 using StoryEditor.UI;
