+++ Assets/Scripts/Editor/UI/BGMEditor.cs
@@ -1,6 +1,7 @@
 using UnityEngine;
 using UnityEditor;
 using AD.Audio;
+using AD.Story;
 using System.Linq;
 using UnityEngine.AddressableAssets;
