+++ Assets/Scripts/Editor/UI/ChoiceEditor.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using UnityEditor;
+using AD.Story;
 using StoryEditor.Controllers;
 using System.Collections.Generic;
