+++ Assets/Scripts/Editor/UI/IStoryEntryEditor.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using AD.Story;
 
 namespace StoryEditor.UI
 {
