+++ Assets/Scripts/Editor/UI/StoryEditorPanel.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using UnityEditor;
+using AD.Story;
 using StoryEditor.Controllers;
 
 namespace StoryEditor.UI
