+++ Assets/Scripts/Editor/UI/StoryEntryEditorFactory.cs
@@ -1,6 +1,7 @@
 using StoryEditor.Controllers;
 using System.Collections.Generic;
 using UnityEngine;
+using AD.Story;
 
 namespace StoryEditor.UI
 {
