+++ Assets/Scripts/Editor/UI/StoryEntryPanel.cs
@@ -1,6 +1,7 @@
 using UnityEngine;
 using UnityEditor;
 using StoryEditor.Controllers;
+using AD.Story;
 
 namespace StoryEditor.UI
 {
