+++ Assets/Scripts/Editor/UI/VFXEditor.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using UnityEditor;
+using AD.Story;
 
 namespace StoryEditor.UI
 {
