+++ Assets/Scripts/Event/Events/Game Events/ChoiceEvent.cs
@@ -78,7 +78,7 @@ public class ChoiceEvent : GameEventBase<ChoiceEventInfo>
         // Get the selected choice and process it
         string selectedChoice = UIController.Instance.GetSelectedChoice();
 
-        StoryController.Instance.ProcessSelectedChoice(selectedChoice);
+        AD.Story.StoryController.Instance.ProcessSelectedChoice(selectedChoice);
 
         TerminateEvent();
     }
