--- Assets/Scripts/GameData/CharacterExpressionAsset.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 277d3d82b1eb84bf98ca4a102a7297ae
