+++ Assets/Scripts/GameData/Story/AssetPath.cs
@@ -2,6 +2,7 @@ namespace AD.Story
 {
     public static class AssetPath
     {
-        public const string BackgroundCG = "GameData/Story/BackgroundCGAsset";
+        public const string BackgroundCG                = "GameData/Story/BackgroundCGAsset";
+        public const string CharacterExpressionAsset    = "GameData/CharacterExpressionAsset";
     }
 }
