+++ Assets/Scripts/GameData/Story/CharacterExpressionAsset.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 251bed377f48dd2428a3b346e7bd178b
