--- Assets/Scripts/Story/Character CG/CharacterCGController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 40e94b2dd0d284abdb3d0cf41a2f8b26
