+++ Assets/Scripts/Story/Character CG/CharacterCGPresentor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8568121b0601ac74a9abfe2a0fd033c1
