+++ Assets/Scripts/Story/IStoryPresenter.cs
@@ -1,8 +1,11 @@
 using System;
 using Cysharp.Threading.Tasks;
 
-public interface IStoryPresenter
+namespace AD.Story
 {
-    void Initialize(StoryController storyController, StoryUIView uiView);
-    UniTask ProgressStoryEntry(StoryEntry storyEntry);
+    public interface IStoryPresenter
+    {
+        void Initialize(StoryController storyController, StoryUIView uiView);
+        UniTask ProgressStoryEntry(StoryEntry storyEntry);
+    }
 }
